+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Tests (test/utils/is.test.js) for all function in src/utils/is.js
 
 ## [2.0.0-alpha.6] - 2018-09-13
 ### Added
@@ -14,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - isObject returned true when passing an array
 
 ### Changed
-- do not ignore package-lock.json anymore in .gitignore 
+- do not ignore package-lock.json anymore in .gitignore
 
 ## [2.0.0-alpha.5] - 2018-09-12
 ### Added