+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests (test/utils/is.test.js) for all function in src/utils/is.js
 - Tests (test/core/createElement/createElement.test.js) for createElement in src/core/createElement/createElement.js
 
+## [2.0.0-alpha.7] - 2018-09-16
+### Changed
+- better typechecking (#17)
+
+### Added
+- additional checks for: window, Date, Error, Function,  Number, RegExp, Undefined (#17)
+
 ## [2.0.0-alpha.6] - 2018-09-13
 ### Added
 - Tests for is.js